Another update for IP address handling.  Changed to non-blocking model of comparing and correcting ID values for players if needed.
Ratio updating is busted and still need to fix it.  Currently commented out.

Also still need to dissaociate client ID values when a player quits so no weirdness happens when we don't have IP addresses/connect logs.

Going to think on solutions for calculating awards as well.  Since IP addresses aren't in the main log file, determining the identity of players during award calculations will be tricky.  One thought is to append the generated ID that we create to each player in each log entry and then do the stats calculations on that modified log set.