Updates to filter so we get only one player packet per player when sent to the frontend.
Places/rank are calculated based upon players on the allies, axis or spectator team.  Places/rank update on kill/death and when a player quits.  This can result in a bit of a desync as the place/rank will only update upon the next player packet being created.  (AKA: I'm not sending a player update packet each time someone's place changes.)

Removed NPC kills from showing up on the map.  When you die to the world or an unmanned vehicle this is shown as a suicide.  This seems reasonable as falling is considered a world-kill as well as getting run over or killed in an explosion by an unmanned vehicle.

Switched JSON libraries as I couldn't get the old one to not put \\ in front of slashes in strings.  This ought to make photos possible.  (can't get my laptop to see my desktop in the network so can't check at the moment)